Update event tests and add workflow
