+++ SdkTestAutomation.Api/ApiUrl.cs
@@ -8,5 +8,10 @@ namespace SdkTestAutomation.Api
         {
             public static string EventUrl => $"{TestConfig.ApiUrl}/event";
         }
+        
+        public static class WorkflowResource
+        {
+            public static string WorkflowUrl => $"{TestConfig.ApiUrl}/workflow";
+        }
     }
 }
