+++ SdkTestAutomation.Api/Conductor/WorkflowResource/Request/GetWorkflowRequest.cs
@@ -1,0 +1,5 @@
+namespace SdkTestAutomation.Api.Conductor.WorkflowResource.Request;
+
+public class GetWorkflowRequest : BaseRequest
+{
+}
