+++ SdkTestAutomation.Api/Conductor/WorkflowResource/Response/GetEventResponse.cs
@@ -1,0 +1,7 @@
+namespace SdkTestAutomation.Api.Conductor.WorkflowResource.Response;
+
+public class GetWorkflowResponse
+{
+    public string Status { get; set; }
+    public string WorkflowId { get; set; }
+}
