+++ SdkTestAutomation.Api/SdkTestAutomation.Api.csproj
@@ -16,6 +16,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <Folder Include="Conductor\WorkflowResource\Models\" />
       <Folder Include="Extensions\" />
     </ItemGroup>
